[PS-42] refactor: 액세스 토큰이 body에 담기도록 변경
