+++ src/main/java/com/tf4/photospot/auth/application/response/LoginResponse.java
@@ -1,5 +1,6 @@
 package com.tf4.photospot.auth.application.response;
 
 public record LoginResponse(
+	String accessToken,
 	boolean hasLoggedInBefore) {
 }
